+++ src/main/java/com/moondysmell/yanoljaclone/exception/ErrorCode.java
@@ -9,9 +9,9 @@ public enum ErrorCode {
     INVALID_CODE(400, "C001", "Invalid Code"),
 
     //1000:Accom
-    NOT_ENOUGH_ROOM(2000,"A002","No Room Anymore");
 
     //2000:Reservation
+    NOT_ENOUGH_ROOM(2000,"A002","No Room Anymore");
 
     //3000: Customer
