+++ src/main/java/com/moondysmell/yanoljaclone/service/ReservationService.java
@@ -3,13 +3,17 @@ package com.moondysmell.yanoljaclone.service;
 import com.moondysmell.yanoljaclone.domain.*;
 import com.moondysmell.yanoljaclone.domain.dto.ReservationMakeDto;
 import com.moondysmell.yanoljaclone.domain.dto.ReservationResponseDto;
+import com.moondysmell.yanoljaclone.exception.NotEnoughRoomException;
 import com.moondysmell.yanoljaclone.repository.AccommodationRepository;
 import com.moondysmell.yanoljaclone.repository.UserRepository;
 import com.moondysmell.yanoljaclone.repository.reservation.ReservationRepository;
 //import com.moondysmell.yanoljaclone.repository.reservation.ReservationRepositoryCustom;
+import java.time.LocalDate;
+import java.time.ZoneId;
 import java.util.Date;
 import java.util.Optional;
 import lombok.*;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,14 +25,17 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 @Transactional(readOnly=true)
 @RequiredArgsConstructor
 public class ReservationService {
     @Autowired
     private final AccommodationRepository accomRepository;
+
     @Autowired
     private final UserRepository userRepository;
+
     @Autowired
     private final ReservationRepository reservRepositoty;
 
@@ -42,10 +49,6 @@ public class ReservationService {
         }
 
         return reservList;
-        //return reserv.stream().map(ReservationResponseDto::new).collect(Collectors.toList());
-        //return reservList.stream()
-         //       .map(reserv -> new ReservationResponseDto(reserv))
-         //       .collect(Collectors.toList());
     }
 
     @Transactional
@@ -54,13 +57,21 @@ public class ReservationService {
         int accomId = reservationMakeDto.getAccomId();
         String userName = reservationMakeDto.getUserName();
         String phoneNum = reservationMakeDto.getPhoneNum();
+        int roomCnt = reservationMakeDto.getRoomCnt();
+        LocalDate checkin = convertDateToLocalDate(reservationMakeDto.getCheckin());
+        LocalDate checkout = convertDateToLocalDate(reservationMakeDto.getCheckout());
 
 
-        Accommodation accom = accomRepository.findById(accomId).orElseThrow(
+        Accommodation targetAccom = accomRepository.findById(accomId).orElseThrow(
                 () -> {
                     throw new IllegalArgumentException("해당 숙박업소를 조회할 수 없습니다.");
                 }
         );
+
+        //예약가능한 날짜와 방
+        int emptyRoomCnt = countEmptyRoomsByAccomIdAndDate(accomId, checkin, checkout);
+        if (emptyRoomCnt < roomCnt) throw new NotEnoughRoomException();
+
 
         //예약내역 입력과 동시에 customerTable에 customer정보 insert
         //입력한 customer객체의 정보를 예약 등록하는 내역에 설정
@@ -73,7 +84,7 @@ public class ReservationService {
 
 
         newReserv.addCustomer(customer);
-        newReserv.setAccom(accom);
+        newReserv.setAccom(targetAccom);
 
         newReserv = reservRepositoty.save(newReserv);
 
@@ -89,5 +100,34 @@ public class ReservationService {
 
 //         return reservations.size();
     }
+
+    public LocalDate convertDateToLocalDate(Date dt) {
+        return dt.toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
+    }
+
+    //reservationService 에서 예약 가능한지 확인할 때 사용하는 함수
+    public int countEmptyRoomsByAccomIdAndDate(int accomId, LocalDate from, LocalDate to) {
+        Optional<Accommodation> targetAccom = accomRepository.findById(accomId);
+        if (targetAccom.isEmpty())
+            throw new RuntimeException("accomId 가 존재하지 않습니다. 다시 확인해주세요");
+
+        int emptyRoomResult = 9999;
+        int totalRoomCnt = targetAccom.get().getRoomCnt();
+
+        // from 부터 to까지 모든 날짜를 다 List로 만들어 놓고 for문 돌림
+        List<LocalDate> listOfDates = from.datesUntil(to)
+                .collect(Collectors.toList());
+        log.info("from과 to의 날짜 차이: " + listOfDates.size());
+
+        for (LocalDate ld : listOfDates) {
+            Date targetDate = java.sql.Date.valueOf(ld);
+            int reservedCnt = getReservByAccomIdAndDate(accomId, targetDate);
+            int emptyCnt = totalRoomCnt - reservedCnt;
+            emptyRoomResult = (emptyRoomResult > emptyCnt) ? emptyCnt : emptyRoomResult;
+        }
+
+        return emptyRoomResult;
+    }
+
 
 }
